백준-오등큰수 문제풀이

+++ BOJ/src/p6000/옥상정원꾸미기_6198.java
@@ -21,22 +21,22 @@ public class 옥상정원꾸미기_6198 {
 		
 		for(int i=0; i<N; i++) { input[i] = Integer.parseInt(br.readLine()); }
 		
-		Stack<long[]> st = new Stack<>();
-		long answer = 0;
+		Stack<int[]> st = new Stack<>();
+		int answer = 0;
 		
 		for(int i=N-1; i>=0; i--) {
 			int num = input[i];
 			if(st.isEmpty()) {
-				st.add(new long[] {num, 0}); // 현재 높이와 cnt 데이터 삽입
+				st.add(new int[] {num, 0}); // 현재 높이와 cnt 데이터 삽입
 				continue;
 			}
 			
-			long cnt = 0;
+			int cnt = 0;
 			while(!st.isEmpty() && num > st.peek()[0]) { // 오른쪽에서 크거나 같은 부분이 나올때까지 pop
 				cnt += (st.pop()[1] + 1);
 			}
 			
-			st.add(new long[] {num, cnt});
+			st.add(new int[] {num, cnt});
 			
 			answer += cnt;
 		}
